Remove delay for geocode adress and keep sequential fetch in seperate file
